AI authored Lissajous curve generator
Code generated from the following prompt:

Display a dialog box with input fields for variables named 'a','b', 'width'. Create an array with a data series from zero to 2 * Pi. Use that data series and the variables 'a' and 'b' in parametric equations for a Lissajous curve. Plot the Lissajous curve as a path in SVG. Turn fill off. Set the stroke to black and the width to the value of 'width'. Set the parameters of the viewBox such that the Lissajous curve appears in the center of the browser window. Append a link to download a file that contains the SVG element. Name the file using a string that includes the current year, month, day concatenated with "_Lissajous_curve.svg"